Move test-logging outside helpers/ directory
